feat: build project using Svelte
This is a complete rewrite of the project that works
with the current AVIF implementation using Svelte
for components reuse.

This project no longer relies on signed URLs as we can
use the s=none method that is now supported for DCR.

Images are downloaded only once for more accurate measurements.

Totals for the whole corpus are given for each configuration.